yolo/requirements.txt
click==8.1.3
opencv-python==4.7.0.72
python-dotenv==1.0.0
ultralytics==8.0.61
IPython==8.12.0